feat: APP-670 merge edit profile with dashboard, update imports and routes, code clean up
