+++ web-marketplace/src/components/organisms/DashboardNavigation/DashboardNavigation.MobileHeader.tsx
@@ -5,7 +5,7 @@ import { Subtitle } from 'web-components/src/components/typography/Subtitle';
 import UserAvatar from 'web-components/src/components/user/UserAvatar';
 import { cn } from 'web-components/src/utils/styles/cn';
 
-import { getDefaultAvatar } from 'pages/ProfileEdit/ProfileEdit.utils';
+import { getDefaultAvatar } from 'pages/Dashboard/Dashboard.utils';
 
 import { UNNAMED } from './DashboardNavigation.constants';
 import { DashboardNavHeaderData } from './DashboardNavigation.types';
