+++ web-marketplace/src/components/organisms/RolesForm/components/RoleField/RoleField.utils.tsx
@@ -6,8 +6,8 @@ import { UseStateSetter } from 'web-components/src/types/react/useState';
 import { AccountFieldsFragment, AccountType, Maybe } from 'generated/graphql';
 import { TranslatorType } from 'lib/i18n/i18n.types';
 
-import { DEFAULT_NAME } from 'pages/ProfileEdit/ProfileEdit.constants';
-import { getDefaultAvatar } from 'pages/ProfileEdit/ProfileEdit.utils';
+import { DEFAULT_NAME } from 'pages/Dashboard/Dashboard.constants';
+import { getDefaultAvatar } from 'pages/Dashboard/Dashboard.utils';
 
 import { ProfileModalSchemaType } from '../ProfileModal/ProfileModal.schema';
 import { AddNewProfile } from './RoleField.AddNewProfile';
