--- web-marketplace/src/pages/Dashboard/Dashboard.styles.tsx
@@ -1,5 +1,0 @@
-import { SxProps } from '@mui/material';
-
-export const dashBoardStyles = {
-  padTop: { pt: 10, pb: [21.25, 28.28] } as SxProps,
-};
