+++ web-marketplace/src/pages/Dashboard/Dashboard.utils.tsx
@@ -4,11 +4,19 @@ import TwitterIcon from 'web-components/src/components/icons/social/TwitterIcon'
 import WebsiteLinkIcon from 'web-components/src/components/icons/social/WebsiteLinkIcon';
 import { SocialLink } from 'web-components/src/components/organisms/ProfileHeader/ProfileHeader.types';
 
-import { Account, AccountByIdQuery } from 'generated/graphql';
+import {
+  Account,
+  AccountByIdQuery,
+  AccountType,
+  Maybe,
+} from 'generated/graphql';
 import { Wallet } from 'lib/wallet/wallet';
 
-import { DEFAULT_PROFILE_BG } from 'pages/ProfileEdit/ProfileEdit.constants';
-import { getDefaultAvatar } from 'pages/ProfileEdit/ProfileEdit.utils';
+import {
+  DEFAULT_PROFILE_BG,
+  DEFAULT_PROFILE_COMPANY_AVATAR,
+  DEFAULT_PROFILE_USER_AVATAR,
+} from './Dashboard.constants';
 
 /* getSocialsLinks */
 
@@ -77,4 +85,18 @@ export const getWalletAddress = ({
   wallet,
 }: GetWalletAddressParams) => {
   return activeAccount ? activeAccount.addr : wallet?.address;
+};
+
+export const getDefaultAvatar = (account?: Maybe<Pick<Account, 'type'>>) => {
+  const isOrganization = account?.type === AccountType.Organization;
+  const defaultAvatar = isOrganization
+    ? DEFAULT_PROFILE_COMPANY_AVATAR
+    : DEFAULT_PROFILE_USER_AVATAR;
+
+  return defaultAvatar;
+};
+export const getProfileUrl = (account: { addr?: string; id?: string }) => {
+  // Use address if available, otherwise fall back to ID
+  const identifier = account.addr || account.id;
+  return `/profiles/${identifier}`;
 };
