--- web-marketplace/src/pages/ProfileEdit/ProfileEdit.tsx
@@ -1,291 +1,0 @@
-import { useEffect, useMemo, useRef, useState } from 'react';
-import { Outlet, useLocation, useNavigate } from 'react-router-dom';
-import { useLingui } from '@lingui/react';
-import { useQuery } from '@tanstack/react-query';
-import { useAtom, useSetAtom } from 'jotai';
-import { startCase } from 'lodash';
-
-import { Flex } from 'web-components/src/components/box';
-import BridgeIcon from 'web-components/src/components/icons/BridgeIcon';
-import CreditsIcon from 'web-components/src/components/icons/CreditsIcon';
-import HamburgerIcon from 'web-components/src/components/icons/HamburgerIcon';
-import { SaveChangesWarningModal } from 'web-components/src/components/modal/SaveChangesWarningModal/SaveChangesWarningModal';
-import { IconTabProps } from 'web-components/src/components/tabs/IconTab';
-import { IconTabs } from 'web-components/src/components/tabs/IconTabs';
-import { Title } from 'web-components/src/components/typography';
-import { cn } from 'web-components/src/utils/styles/cn';
-
-import { selectedLanguageAtom } from 'lib/atoms/languageSwitcher.atoms';
-import { isProfileEditDirtyRef } from 'lib/atoms/ref.atoms';
-import { useAuth } from 'lib/auth/auth';
-import { client as sanityClient } from 'lib/clients/sanity';
-import { DISCARD_CHANGES_TITLE } from 'lib/constants/shared.constants';
-import { getAllProfilePageQuery } from 'lib/queries/react-query/sanity/getAllProfilePageQuery/getAllProfilePageQuery';
-import { useWallet } from 'lib/wallet/wallet';
-
-import { useProfileItems } from 'pages/Dashboard/hooks/useProfileItems';
-import { Link } from 'components/atoms';
-import WithLoader from 'components/atoms/WithLoader';
-import { DashboardNavigation } from 'components/organisms/DashboardNavigation';
-import { DashboardNavigationMobileHeader } from 'components/organisms/DashboardNavigation/DashboardNavigation.MobileHeader';
-
-import {
-  DISCARD_CHANGES_BODY,
-  DISCARD_CHANGES_BUTTON,
-} from '../../lib/constants/shared.constants';
-import { usePathSection } from './hooks/usePathSection';
-import { BRIDGE, PORTFOLIO } from './ProfileEdit.constants';
-import { ViewProfileButton } from './ProfileEdit.ViewProfile';
-
-export const ProfileEdit = () => {
-  const { _ } = useLingui();
-  const [selectedLanguage] = useAtom(selectedLanguageAtom);
-  const { accountChanging, loginDisabled } = useWallet();
-  const { loading, activeAccount, authenticatedAccounts } = useAuth();
-
-  const [isWarningModalOpen, setIsWarningModalOpen] = useState<
-    string | undefined
-  >(undefined);
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-  const setIsProfileEditDirtyref = useSetAtom(isProfileEditDirtyRef);
-  const isDirtyRef = useRef<boolean>(false);
-  const navigate = useNavigate();
-  const { pathname } = useLocation();
-  const section = usePathSection();
-  const { isCreditClassCreator, isProjectAdmin, isIssuer } = useProfileItems(
-    {},
-  );
-
-  const { data: sanityProfilePageData } = useQuery(
-    getAllProfilePageQuery({
-      sanityClient,
-      enabled: !!sanityClient,
-      languageCode: selectedLanguage,
-    }),
-  );
-
-  const onNavClick = (sectionName: string): void => {
-    const isFormDirty = isDirtyRef.current;
-    const path = `/dashboard/${sectionName.replace(' ', '-')}`;
-    if (isFormDirty) {
-      setIsWarningModalOpen(path);
-    } else {
-      navigate(path);
-    }
-  };
-
-  useEffect(() => {
-    if (isDirtyRef && setIsProfileEditDirtyref) {
-      setIsProfileEditDirtyref(isDirtyRef);
-    }
-  }, [isDirtyRef, setIsProfileEditDirtyref]);
-
-  function onAccountSelect(id: string) {
-    throw new Error('Function not implemented.');
-  }
-
-  const dashboardContextValue = useMemo(
-    () => ({
-      isCreditClassCreator,
-      isProjectAdmin,
-      isIssuer,
-      sanityProfilePageData,
-    }),
-    [isCreditClassCreator, isProjectAdmin, isIssuer, sanityProfilePageData],
-  );
-
-  const portfolioTabs: IconTabProps[] = useMemo(
-    () => [
-      {
-        label: _(PORTFOLIO),
-        href: '/dashboard/portfolio',
-        icon: <CreditsIcon fontSize="small" linearGradient />,
-      },
-      {
-        label: _(BRIDGE),
-        href: '/dashboard/portfolio/bridge',
-        icon: <BridgeIcon linearGradient />,
-      },
-    ],
-    [_],
-  );
-
-  const activePortfolioTab = Math.max(
-    portfolioTabs.findIndex(tab => {
-      if (tab.href === '/dashboard/portfolio/bridge') {
-        return pathname.includes('/dashboard/portfolio/bridge');
-      }
-      if (tab.href === '/dashboard/portfolio') {
-        return pathname === '/dashboard/portfolio';
-      }
-      return false;
-    }),
-    0,
-  );
-
-  return (
-    <div className="bg-grey-100">
-      <div className="relative md:flex md:min-h-screen">
-        {/* Mobile Header */}
-        {activeAccount && (
-          <DashboardNavigationMobileHeader
-            activeAccount={{
-              ...activeAccount,
-              address: activeAccount.addr ?? '',
-              type: activeAccount.type === 'USER' ? 'user' : 'org',
-              image: activeAccount.image || '',
-            }}
-            onMenuClick={() => setMobileMenuOpen(true)}
-            mobileMenuOpen={mobileMenuOpen}
-          />
-        )}
-
-        {/* Mobile overlay with blur */}
-        {mobileMenuOpen && (
-          <div
-            className="md:hidden fixed inset-0 bg-black bg-opacity-50 backdrop-blur-sm z-30"
-            onClick={() => setMobileMenuOpen(false)}
-          />
-        )}
-
-        {/* Left sidebar navigation */}
-        {activeAccount && (
-          <div className="md:sticky md:top-0 md:h-screen">
-            <DashboardNavigation
-              currentPath={pathname}
-              onNavItemClick={onNavClick}
-              isIssuer={isIssuer}
-              mobileMenuOpen={mobileMenuOpen}
-              onLogout={() => {
-                setIsWarningModalOpen(undefined);
-                navigate('/logout');
-                setMobileMenuOpen(false);
-              }}
-              onCloseMobile={() => setMobileMenuOpen(false)}
-              onExitClick={() => {
-                setIsWarningModalOpen(undefined);
-                navigate('/');
-                setMobileMenuOpen(false);
-              }}
-              header={{
-                activeAccount: {
-                  ...activeAccount,
-                  address: activeAccount.addr ?? '',
-                  type: activeAccount.type === 'USER' ? 'user' : 'org',
-                  image: activeAccount.image || '',
-                },
-                accounts: (authenticatedAccounts || []).map(account => ({
-                  id: account?.id,
-                  name: account?.name || '',
-                  address: account?.addr || '',
-                  type: account?.type === 'USER' ? 'user' : 'org',
-                  image: account?.image || '',
-                })),
-                onAccountSelect: (id: string) => {
-                  setIsWarningModalOpen(undefined);
-                  onAccountSelect(id);
-                  setMobileMenuOpen(false);
-                },
-                onViewProfileClick: (path: string) => {
-                  setIsWarningModalOpen(undefined);
-                  navigate(`${path}`);
-                  setMobileMenuOpen(false);
-                },
-              }}
-            />
-          </div>
-        )}
-
-        {/* Content area - add blur when mobile menu is open */}
-        <div
-          className={cn(
-            'flex-1 min-w-0 w-full md:w-auto pt-[60px] md:pt-0 transition-all duration-300',
-            mobileMenuOpen && 'md:blur-0 blur-sm',
-          )}
-        >
-          <div
-            className={cn(
-              'px-10 py-25 md:py-40 md:px-30 2xl:mx-auto',
-              section === 'profile' || section === 'settings'
-                ? 'max-w-[767px]'
-                : 'max-w-[1400px]',
-            )}
-          >
-            <Flex
-              sx={{
-                flexDirection: 'column',
-                alignItems: 'center',
-              }}
-              className={cn('w-full', section ? 'flex' : 'hidden')}
-            >
-              <Flex
-                justifyContent="space-between"
-                alignContent="center"
-                className="w-full h-50"
-              >
-                <Title
-                  variant="h1"
-                  className="text-[21px] md:text-[32px] leading-[1.4]"
-                >
-                  {pathname.includes('/portfolio')
-                    ? 'Portfolio'
-                    : startCase(section)}
-                </Title>
-                {(section === 'credit-classes' ||
-                  section === 'projects' ||
-                  section === 'portfolio' ||
-                  section === 'profile') && (
-                  <ViewProfileButton
-                    setIsWarningModalOpen={setIsWarningModalOpen}
-                    section={section}
-                    activeAccount={
-                      activeAccount
-                        ? {
-                            ...activeAccount,
-                            addr: activeAccount.addr ?? undefined,
-                          }
-                        : undefined
-                    }
-                  />
-                )}
-              </Flex>
-
-              {(section === 'portfolio' ||
-                pathname.includes('/portfolio/bridge')) && (
-                <div className="w-full mb-8">
-                  <IconTabs
-                    aria-label="portfolio tabs"
-                    tabs={portfolioTabs}
-                    activeTab={activePortfolioTab}
-                    linkComponent={Link}
-                    mobileFullWidth
-                  />
-                </div>
-              )}
-
-              <WithLoader isLoading={accountChanging || loading}>
-                <div className="rounded-md border border-grey-200 bg-grey-0 lg:mt-30 w-full">
-                  <Outlet context={dashboardContextValue} />
-                </div>
-              </WithLoader>
-            </Flex>
-          </div>
-        </div>
-      </div>
-
-      <SaveChangesWarningModal
-        open={!!isWarningModalOpen}
-        title={_(DISCARD_CHANGES_TITLE)}
-        bodyText={_(DISCARD_CHANGES_BODY)}
-        buttonText={_(DISCARD_CHANGES_BUTTON)}
-        navigate={() => {
-          if (isWarningModalOpen) navigate(isWarningModalOpen);
-          isDirtyRef.current = false;
-        }}
-        onClose={() => {
-          setIsWarningModalOpen(undefined);
-        }}
-      />
-    </div>
-  );
-};
