--- web-marketplace/src/pages/ProfileEdit/index.ts
@@ -1,1 +1,0 @@
-export { ProfileEdit as default } from './ProfileEdit';
